Authorization: Deleting a User's Task

+++ src/tasks/tasks.service.ts
@@ -14,12 +14,12 @@ export class TasksService {
         private taskRepository: TaskRepository,
     ) { }
 
-    async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]>{
+    async getTasks(filterDto: GetTaskFilterDto, user: User): Promise<Task[]> {
         return this.taskRepository.getTasks(filterDto, user);
     }
 
     async getTaskById(id: number, user: User): Promise<Task> {
-        const found = await this.taskRepository.findOne({ where: {id, userId: user.id}});
+        const found = await this.taskRepository.findOne({ where: { id, userId: user.id } });
 
         if (!found) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
@@ -32,8 +32,8 @@ export class TasksService {
         return this.taskRepository.createTask(createTaskDto, user);
     }
 
-    async deleteTask(id: number): Promise<void> {
-        const result = await this.taskRepository.delete(id);
+    async deleteTask(id: number, user: User): Promise<void> {
+        const result = await this.taskRepository.delete({ id, userId: user.id });
 
         if (result.affected == 0) {
             throw new NotFoundException(`Task with ID "${id}" not found`);
